refactor: re-arranging files for better structure
