+++ events/interaction/handlers/button.js
@@ -1,4 +1,4 @@
-import { handleCreateUserButton } from "./features/users/create-user-modal.js";
+import { handleCreateUserButton } from "../features/users/create-user-modal.js";
 
 export const buttonHandlers = {
   create_user_modal_button: handleCreateUserButton,
