+++ events/interaction/handlers/handleEvent.js
