+++ events/interaction/handlers/modal.js
@@ -1,4 +1,4 @@
-import { handleCreateUserModal } from "./features/users/handle-user-modal.js";
+import { handleCreateUserModal } from "../features/users/handle-user-modal.js";
 
 export const modalSubmitHandlers = {
   create_user_modal: handleCreateUserModal,
