+++ events/interaction/handlers/select.js
